S3File: use dedicated temp stream to get size
In order to avoid prematurely creating and using our persistent
connection which will be used by the `ZipStream` library to actually
fetch the file contents, we here switch to a separately and immediately
discarded stream to use to fetch the file size (which cannot impact the
main content-fetching stream).

This solution will not break the fix to the issue of filename parsing
that #121 resolved.

Test results:

```
Runtime:       PHP 8.3.11
Configuration: /home/zmd/Code/laravel-zipstream/phpunit.xml.dist

...................                                               19 / 19 (100%)

Time: 00:03.184, Memory: 44.00 MB
```

Fixes #125.